feat(checkbox): convert to typescript
